move tests related to functions in _matplotlib_backend to test_matplotlib_backend
